attached_assets/requirements_1752143555494.txt
numpy
scipy
matplotlib
sounddevice
soundfile
openai-whisper
torch
ffmpeg-python
